Create Basic machine learning problem.ipynb
I loaded a music dataset from a CSV file and separated it into features and the target variable, with the features being all columns except the 'genre' column and the target being the 'genre' column. I then created a Decision Tree Classifier model and trained it on the features and target to learn the relationship between music attributes and their genres. Finally, I used the trained model to predict the genres of two new samples with specific attribute values.